Use AM_ICONV
AC_SEARCH_LIBS does not detect GNU libiconv because its symbols are
exported in the "libiconv" namespace instead of "iconv".  The AM_ICONV
macro correctly detects both glibc and GNU libiconv, defines HAVE_ICONV,
ICONV_CONST, and LIBICONV depending on the system.  The config.rpath
file is required by this macro.

This adds a dependency on the aclocal macros from gettext (gettext-devel
in some distros) when building from git, but not when building from a
tarball.

Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>
Reviewed-by: Gaetan Nadon <memsize@videotron.ca>