Allow sub-second updates.
Lets you run the clock at whatever tick rate you like.

Note that this changes the behaviour of the clock when the update
isn't a divisor of the number of seconds in a day. In the old code,
the time to display was an integer number of intervals from the start
of the year, in the new code, it is an integer number of intervals
from the start of the day.

I like that better as it makes the seconds consistent from day to day,
even if it means an uneven tick at local midnight. Given that I
suspect no-one has ever done this in practice, I doubt it really
matters.

At least the code is simpler now, while supporting updates at the
level of the Xt timers, which are milliseconds.

Signed-off-by: Keith Packard <keithp@keithp.com>