Configurable shapes and AM/PM marker
Allow configuring the shape of the hands and scale ticks by describing
a series of triangles.

Add AM/PM markers with configurable color and shape. Add flags for
whether to rotate the AM/PM markers along the hour hand and whether to
render them over or under the hour hand.